@@ -1,16 +1,18 @@
 <?php
 namespace Datatrics\API\Modules;
 
+use Datatrics\API\Client;
+
 class Tracker extends Base
 {
     /**
      * Private constructor so only the client can create this
-     * @param string $apikey
-     * @param string $projectid
+     * @param Client $client
      */
-    public function __construct($apikey, $projectid)
+    public function __construct(Client $client)
     {
-        parent::__construct($apikey, "/project/" . $projectid . "/tracker");
+        parent::__construct($client);
+        $this->SetUrl("/project/" . $this->GetClient()->GetProjectId() . "/tracker");
     }
 
     /**
@@ -18,8 +20,8 @@ class Tracker extends Base
      * @param object Containing query arguments
      * @return object Result of the request
      */
-    public function Stats($args = array())
+    public function Stats($args = [])
     {
-        $this->request(self::HTTP_GET, "/stats?".http_build_query($args));
+        $this->GetClient()->Get($this->GetUrl()."/stats", $args);
     }
 }
