@@ -1,6 +1,7 @@
 <?php
 namespace Datatrics\API;
 
+use GuzzleHttp\Client as HttpClient;
 use Datatrics\API\Modules\Apikey;
 use Datatrics\API\Modules\Behavior;
 use Datatrics\API\Modules\Box;
@@ -28,28 +29,61 @@ use Datatrics\API\Modules\Tric;
 use Datatrics\API\Modules\Trigger;
 use Datatrics\API\Modules\User;
 use Datatrics\API\Modules\Webhook;
+use GuzzleHttp\Psr7\Request;
 
 class Client
 {
     /**
-     * Version of our client.
+     * @const Version of our client.
      */
-    const CLIENT_VERSION = '2.0.0';
+    const CLIENT_VERSION = '2.0';
+
+    /**
+     * @const HTTP Method GET
+     */
+    const HTTP_GET = 'GET';
+
+    /**
+     * @const HTTP Method POST
+     */
+    const HTTP_POST = 'POST';
+
+    /**
+     * @const HTTP Method PUT
+     */
+    const HTTP_PUT = 'PUT';
+
+    /**
+     * @const HTTP Method DELETE
+     */
+    const HTTP_DELETE = 'DELETE';
 
     /**
      * Version of the remote API.
+     *
+     * @var string
      */
-    protected $api_version = '2.0';
+    private $_api_version = '2.0';
 
     /**
      * @var string
      */
-    protected $api_key;
+    private $_api_endpoint = 'https://api.datatrics.com';
 
     /**
      * @var string
      */
-    private $projectId;
+    private $_api_key;
+
+    /**
+     * @var HttpClient
+     */
+    private $_http_client;
+
+    /**
+     * @var string
+     */
+    private $_projectId;
 
     /**
      * @var Apikey
@@ -188,92 +222,281 @@ class Client
 
     /**
      * Create a new API instance
-     * @param string $apiKey
-     * @param string $projectId
+     *
+     * @param string $apiKey    The API key
+     * @param string $projectId The Project id
      */
-    public function __construct($apiKey, $projectId)
+    public function __construct($apiKey, $projectId = null)
     {
         $this->SetApiKey($apiKey);
         $this->SetProjectId($projectId);
-
-        $this->registerModules();
+        $this->SetHttpClient();
+        $this->_RegisterModules();
     }
 
     /**
+     * Get the current API version
+     *
      * @return string $api_version
      */
     public function GetApiVersion()
     {
-        return $this->api_version;
+        return $this->_api_version;
     }
 
     /**
+     * Get the API endpoint
+     *
+     * @return string
+     */
+    public function GetApiEndpoint()
+    {
+        return $this->_api_endpoint;
+    }
+
+    /**
+     * Set the API endpoint
+     *
+     * @param string $api_endpoint
+     * @return  Client
+     */
+    public function SetApiEndpoint($api_endpoint)
+    {
+        $this->_api_endpoint = $api_endpoint;
+        $this->SetHttpClient();
+        $this->_RegisterModules();
+        return $this;
+    }
+
+    /**
+     * Get the API key
+     *
      * @return string $api_key
      */
     public function GetApiKey()
     {
-        return $this->api_key;
+        return $this->_api_key;
     }
 
     /**
+     * Get the API key
+     *
      * @param string $api_key
+     * @return Client
      */
     public function SetApiKey($api_key)
     {
-        $this->api_key = $api_key;
-
-        $this->registerModules();
+        $this->_api_key = $api_key;
+        $this->SetHttpClient();
+        $this->_RegisterModules();
+        return $this;
     }
 
     /**
+     * Get the API project id
+     *
      * @return string $projectId
      */
     public function GetProjectId()
     {
-        return $this->projectId;
+        return $this->_projectId;
     }
 
     /**
+     * Set the API project id
+     *
      * @param string $projectId
+     * @return Client
      */
     public function SetProjectId($projectId)
     {
-        $this->projectId = $projectId;
-
-        $this->registerModules();
+        $this->_projectId = $projectId;
+        $this->SetHttpClient();
+        $this->_RegisterModules();
+        return $this;
     }
 
     /**
-     * Register modules
+     * Register Modules
+     *
+     * @return void
      */
-    private function registerModules()
+    private function _RegisterModules()
     {
-        $this->Apikey = new Apikey($this->api_key);
-        $this->Behavior = new Behavior($this->api_key, $this->projectId);
-        $this->Box = new Box($this->api_key, $this->projectId);
-        $this->Bucket = new Bucket($this->api_key, $this->projectId);
-        $this->Campaign = new Campaign($this->api_key, $this->projectId);
-        $this->Card = new Card($this->api_key, $this->projectId);
-        $this->Channel = new Channel($this->api_key, $this->projectId);
-        $this->Content = new Content($this->api_key, $this->projectId);
-        $this->Geo = new Geo($this->api_key);
-        $this->Goal = new Goal($this->api_key, $this->projectId);
-        $this->Interaction = new Interaction($this->api_key, $this->projectId);
-        $this->Journey = new Journey($this->api_key, $this->projectId);
-        $this->Link = new Link($this->api_key, $this->projectId);
-        $this->NextBestAction = new NextBestAction($this->api_key, $this->projectId);
-        $this->Profile = new Profile($this->api_key, $this->projectId);
-        $this->Project = new Project($this->api_key);
-        $this->Sale = new Sale($this->api_key, $this->projectId);
-        $this->Scorecard = new Scorecard($this->api_key, $this->projectId);
-        $this->Segment = new Segment($this->api_key, $this->projectId);
-        $this->Template = new Template($this->api_key, $this->projectId);
-        $this->Theme = new Theme($this->api_key, $this->projectId);
-        $this->Touchpoint = new Touchpoint($this->api_key, $this->projectId);
-        $this->Tracker = new Tracker($this->api_key, $this->projectId);
-        $this->Tric = new Tric($this->api_key, $this->projectId);
-        $this->Trigger = new Trigger($this->api_key, $this->projectId);
-        $this->User = new User($this->api_key, $this->projectId);
-        $this->Webhook = new Webhook($this->api_key, $this->projectId);
+        $this->Apikey = new Apikey($this);
+        $this->Behavior = new Behavior($this);
+        $this->Box = new Box($this);
+        $this->Bucket = new Bucket($this);
+        $this->Campaign = new Campaign($this);
+        $this->Card = new Card($this);
+        $this->Channel = new Channel($this);
+        $this->Content = new Content($this);
+        $this->Geo = new Geo($this);
+        $this->Goal = new Goal($this);
+        $this->Interaction = new Interaction($this);
+        $this->Journey = new Journey($this);
+        $this->Link = new Link($this);
+        $this->NextBestAction = new NextBestAction($this);
+        $this->Profile = new Profile($this);
+        $this->Project = new Project($this);
+        $this->Sale = new Sale($this);
+        $this->Scorecard = new Scorecard($this);
+        $this->Segment = new Segment($this);
+        $this->Template = new Template($this);
+        $this->Theme = new Theme($this);
+        $this->Touchpoint = new Touchpoint($this);
+        $this->Tracker = new Tracker($this);
+        $this->Tric = new Tric($this);
+        $this->Trigger = new Trigger($this);
+        $this->User = new User($this);
+        $this->Webhook = new Webhook($this);
+    }
+
+    /**
+     * Setup the HTTP Client
+     *
+     * @return Client
+     */
+    private function SetHttpClient()
+    {
+        $config = [
+            'base-uri' => $this->GetApiEndpoint(),
+            'headers' => $this->_GetHttpClientHeaders()
+        ];
+        $this->_http_client = new HttpClient($config);
+        return $this;
+    }
+
+    /**
+     * @return HttpClient
+     */
+    private function GetHttpClient()
+    {
+        return $this->_http_client;
+    }
+
+    /**
+     * Define the HTTP headers
+     *
+     * @return array
+     */
+    private function _GetHttpClientHeaders()
+    {
+        $user_agent = 'Datatrics/API '.self::CLIENT_VERSION;
+        return [
+            'accept' => 'application/json',
+            'content-type' => 'application/json',
+            'user-agent' => $user_agent,
+            'x-apikey' => $this->GetApiKey(),
+            'x-client-name' => $user_agent,
+            'x-datatrics-client-info' => php_uname()
+        ];
+    }
+
+    /**
+     * @throws \Exception
+     * @return boolean
+     */
+    public function CheckApiKey()
+    {
+        if (empty($this->GetApiKey())) {
+            throw new \Exception('You have not set an api key. Please use setApiKey() to set the API key.');
+        }
+        return true;
+    }
+
+    /**
+     * @param $url
+     * @param null|array $payload
+     * @return string
+     */
+    public function GetUrl($url, $payload = [])
+    {
+        $url = $this->GetApiEndpoint()."/".$this->GetApiVersion().$url;
+        if (count($payload)) {
+            $url .= "?".http_build_query($payload);
+        }
+        return $url;
+    }
+
+    /**
+     * @param string $method    HTTP Method
+     * @param string $url       The url
+     * @param array $payload    The Payload
+     * @return Request
+     */
+    public function BuildRequest($method, $url, $payload = [])
+    {
+        $body = null;
+        if ($method == self::HTTP_GET) {
+            $url = $this->GetUrl($url, $payload);
+        } elseif ($method == self::HTTP_DELETE) {
+            $url = $this->GetUrl($url, $payload);
+        } else {
+            $url = $this->GetUrl($url);
+            if (count($payload)) {
+                $body = json_encode($payload);
+            }
+        }
+        return new Request($method, $url, $this->_GetHttpClientHeaders(), $body);
+    }
+
+    /**
+     * @param string $method    HTTP Method
+     * @param string $url       The url
+     * @param array $payload    The Payload
+     * @return mixed
+     * @throws \Exception
+     */
+    public function SendRequest($method, $url, $payload = [])
+    {
+        $this->CheckApiKey();
+        $request = $this->BuildRequest($method, $url, $payload);
+        try {
+            $response = $this->GetHttpClient()->send($request);
+        } catch (\Exception $e) {
+            throw $e;
+        }
+        return json_decode($response->getBody(), true);
+    }
+
+    /**
+     * @param string $url
+     * @param array $payload
+     * @return mixed
+     */
+    public function Post($url, $payload = [])
+    {
+        return $this->SendRequest(self::HTTP_POST, $url, $payload);
+    }
+
+    /**
+     * @param string $url
+     * @param array $payload
+     * @return mixed
+     */
+    public function Get($url, $payload = [])
+    {
+        return $this->SendRequest(self::HTTP_GET, $url, $payload);
+    }
+
+    /**
+     * @param string $url
+     * @param array $payload
+     * @return mixed
+     */
+    public function Put($url, $payload = [])
+    {
+        return $this->SendRequest(self::HTTP_PUT, $url, $payload);
+    }
+
+    /**
+     * @param string $url
+     * @param array $payload
+     * @return mixed
+     */
+    public function Delete($url, $payload = [])
+    {
+        return $this->SendRequest(self::HTTP_DELETE, $url, $payload);
     }
 }
