@@ -1,16 +1,18 @@
 <?php
 namespace Datatrics\API\Modules;
 
+use Datatrics\API\Client;
+
 class Behavior extends Base
 {
     /**
      * Private constructor so only the client can create this
-     * @param string $apikey
-     * @param string $projectid
+     * @param Client $client
      */
-    public function __construct($apikey, $projectid)
+    public function __construct(Client $client)
     {
-        parent::__construct($apikey, "/project/" . $projectid . "/behavior");
+        parent::__construct($client);
+        $this->SetUrl("/project/" . $this->GetClient()->GetProjectId() . "/behavior");
     }
 
     /**
@@ -20,7 +22,7 @@ class Behavior extends Base
      */
     public function Get($args = array("limit" => 50))
     {
-        return $this->request(self::HTTP_GET, "?".http_build_query($args));
+        return $this->request(Client::HTTP_GET, $this->GetUrl(), $args);
     }
 
     /**
@@ -30,7 +32,7 @@ class Behavior extends Base
      */
     public function GetVisit($args = array("limit" => 50))
     {
-        return $this->request(self::HTTP_GET, "/visit?".http_build_query($args));
+        return $this->request(Client::HTTP_GET, $this->GetUrl()."/visit", $args);
     }
 
     /**
@@ -41,7 +43,10 @@ class Behavior extends Base
      */
     public function GetEvent($eventId = null, $args = array("limit" => 50))
     {
-        return $eventId == null ? $this->request(self::HTTP_GET, "/event?".http_build_query($args)) : $this->request(self::HTTP_GET, "/event/".$eventId."?".http_build_query($args));
+        if (is_null($eventId)) {
+            return $this->request(Client::HTTP_GET, $this->GetUrl()."/event", $args);
+        }
+        return $this->request(Client::HTTP_GET, $this->GetUrl()."/event/".$eventId, $args);
     }
 
     /**
@@ -51,7 +56,7 @@ class Behavior extends Base
      */
     public function CreateEvent($event)
     {
-        return $this->request(self::HTTP_POST, "/evet", $event);
+        return $this->request(Client::HTTP_POST, $this->GetUrl()."/event", $event);
     }
 
     /**
@@ -66,7 +71,7 @@ class Behavior extends Base
             throw new \Exception("event must contain a eventid");
         }
 
-        return $this->request(self::HTTP_PUT, "/event/".$event['eventid'], $event);
+        return $this->request(Client::HTTP_PUT, $this->GetUrl()."/event/".$event['eventid'], $event);
     }
 
     /**
@@ -75,12 +80,11 @@ class Behavior extends Base
      * @throws \Exception When more that 50 events are provided
      * @return object Result of the request
      */
-    public function UpdateBulk($events)
+    public function Bulk($events)
     {
         if (count($events) > 50) {
             throw new \Exception("Maximum of 50 events allowed at a time");
         }
-
-        return $this->request(self::HTTP_POST, "/event/bulk", $events);
+        return $this->request(Client::HTTP_POST, $this->GetUrl()."/event/bulk", ['items' => $events]);
     }
 }
