@@ -1,15 +1,18 @@
 <?php
 namespace Datatrics\API\Modules;
 
+use Datatrics\API\Client;
+
 class Geo extends Base
 {
     /**
      * Private constructor so only the client can create this
-     * @param string $apikey
+     * @param Client $client
      */
-    public function __construct($apikey)
+    public function __construct(Client $client)
     {
-        parent::__construct($apikey, "/project/geo");
+        parent::__construct($client);
+        $this->SetUrl("/geo");
     }
 
     /**
@@ -20,6 +23,6 @@ class Geo extends Base
     public function Get($ip)
     {
         $args = ['ip' => $ip];
-        $this->request(self::HTTP_GET, "?".http_build_query($args));
+        return $this->GetClient()->Get($this->GetUrl(), $args);
     }
 }
